[Vulkan] Store entries in a separate queue after resetting query pool
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89423

We want to avoid tossing shader log entries when we reset the query pool so that the old entires can be used by the profiler after gathering all profiling data is done.

```get_shader_name_and_execution_duration_ns``` is used for accessing shader names/durations after they are flushed. It will be used with the torch profiler.
ghstack-source-id: 174307259

Differential Revision: [D40119621](https://our.internmc.facebook.com/intern/diff/D40119621/)