Add a way to visualize memory snapshot traces
This adds a d3-based interactive visualization for exploring the memory
allocation traces that the caching allocator can capture. This visualization
code can also be attached to kineto trace information in the future to also
provide visualization for the memory events captured there, which come with
addition information about the graph.

ghstack-source-id: cfa5376ec2ad42c52f68801f47afaf376912e0ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90348